ci: patch haskell-ci workflow to upload exe and call check-advisories
Provide a patch to haskell-ci.yml, which haskell-ci will apply when
regenerating the haskell-ci workflow.  This patch adds steps to
upload the hsec-tools executable as an artifact, and invoke the
check-advisories reusable workflow (which was added in a previous
commit).

There is one notable caveat to this approach.  We use a conditional
based on the `matrix.compiler` variable to ensure that we only
upload one variant of the executable.  Currently, ghc-9.6.1.  The
patch file will need to be manually updated as the build matrix
evolves over time (for example, when moving from 9.6.1 to 9.6.2).

Fixes: https://github.com/haskell/security-advisories/issues/10