tools(check): check filename matches advisory ID
We expect advisories in files name `HSEC-YYYY-NNNN.md`.  The
advisory ID also appears in the TOML metadata.  Add a check that the
ID is the same in both places.

If the filename does not match the expected pattern, we allow it.
The heuristic is ``"HSEC-" `isPrefixOf```, to allow for files that
do not match the pattern (for instance, example advisories).  This
is quite a crude heuristic, and should probably be improved
somewhat, to catch more subtle errors.  But this will do for now.

Fixes: https://github.com/haskell/security-advisories/issues/40