lib/advisories: support multiple references
Update the `Advisory` type, removing the `advisoryUrl` field and
adding `advisoryReferences :: [Reference]`.  Also updating existing
advisories and the examples.

Zero references is valid.  I think that's acceptable.  But we may
wish to update to tooling to warn/notify on missing references.

We also ought to change the URL to a more proper URL type and
enforce (per OSV schema) that it is fully qualified with scheme.
But that refinement can be done later.

Fixes: https://github.com/haskell/security-advisories/issues/37