lib/advisories: add support for out-of-band attributes
We plan to derive the publication and modification dates of
advisories from the Git commit history.  So we need a way to pass
those values to the parser.  Add the `OutOfBandAttributes` type for
this purpose.  Also add the `AttributeOverridePolicy` type, which
controls the behaviour when an attribute is specified out-of-band
*and* in the TOML.  Also add parser helpers, and update the parser
to use these helpers to consider out-of-band values for the
`advisoryPublished` and `advisoryModified` fields.

This commit does not modify the executable to set any out-of-band
attributes.  It provides the framework, and later commits will use
it.