lib/advisories: read timestamps from Git history
Update hsec-tools to read the publication and modification
timestamps from the Git history.  The heuristic uses the commit time
of the first appearance of the file as the publication date, and the
most recent commit that changed the file as the modification date.
Rewrites are followed at Git's default similarity threshold.

The behaviour to query the Git repository is provided in the new
module `Security.Advisories.Git`.

Failures running Git or parsing the output are ignored (e.g. if the
given file is not in a Git repository).  This supports use of the
tool outside the Git context.  The caveat is that the advisory TOML
header must have explicit `"date"` and (if different from `"date"`),
`"modified"` fields.

Also update existing advisories to remove the `"date"` field.

Fixes: https://github.com/haskell/security-advisories/issues/46