HSEC-2023-0003: update versions and references
Refine the type of an existing reference, and add a FIX reference.

Update the affected versions.  The fix version is actually v0.11.2,
and the introduced version is v0.5 or *possibly earlier*.  A big
refactor happened before v0.5 and I'm unsure if previous versions
were affected.  In any case, v0.5 is the earliest version on
Hackage.