ci: publish OSV data to dedicated branch
The osv.dev project will be enhanced to ingest OSV data from our
advisory database.  One of the convenient formats for them is to
track a Git branch containing the raw OSV data.

To facilitate this, add a CI step that exports the OSV data to the
`generated/osv-export` branch.  This step only runs upon a push to
`main`.  It exports all the advisories as OSV JSON files, sorted
into directories by year.  Then it snapshots this content onto the
`generated/osv-export` branch and pushes it.  The commit message
contains a UTC timestamp, and the hash of the corresponding commit
on `main`.

No commit is made unless the OSV content has changed.  The
`generated/osv-export` target branch must already exist.

The GitHub Actions token needs write permissions in the repository,
so that it can push the updated OSV data branch.