lib/advisories: add function to get Git repo root
We want to expand the capabilities of `hsec-tools` so that the tool
can be used to:

- reserve advisories
- prepare new advisory templates
- assign IDs to placeholder IDs (e.g. in CI)
- "fsck" the database

For tolerable(?) user experience, it should be possible to run the
tool anywhere within the *security-advisories* repository, or point
it to (somewhere within) the repository.  To help with this, add:

```haskell
getRepoRoot :: FilePath -> IO (Either GitError FilePath)
```

CLI commands can apply this function to the current working
directory, or a user-specified argument, to find the root of the
repository.