lib/advisories: add filesystem helpers
Add the `Security.Advisories.Filesystem` module with helpers to:

- check that a directory appears to be the *security-advisories*
  repo
- list reserved, in-use or all allocated HSEC IDs
- return the greatest allocated ID in the database
- iterate over reserved IDs, or in-use advisory IDs

Also export values for common path components so that we do not need
to hardcoded them in other places (with associated risk of typos).

These helpers will support the creation of CLI commands that perform
checks or modify the *security-advisories* database.