tools: add reserve command
Add the `hsec-tools reserve` command, which reserves an HSEC ID.

The `--id-mode` option specifies whether to use a placeholder ID, or
allocate the next available ID.  This option is currently required,
but we might change this in the future as we enhance our tooling and
PR automation.

The command operates in the context of the current working
directory.  That is, determines that it is running within an
hsec-tools Git clone, and then creates the file in the appropriate
place.  The user can override the repo location via an optional
positional argument.

Also create the directory in the repo (via the `.gitkeep` idiom).