hsec-tools: include affected FilePath with parse errors
The `listAdvisories` error type is `[ParseAdvisoryError]`.  The
error values do not have context indicating the file that caused the
error.  As a result, the error message(s) do not point to the
problematic file(s).

Update the error type to `[(FilePath, ParseAdvisoryError)]`; each
parse error is paired with the path to the problematic file.  As a
result, user error messages are much more useful.